Correct Entry widget reference
Made a deep copy of the original dictionary so on the 2nd, 3rd, etc. iteration of the dictionary keys we aren't referencing old keys because they're being removed from the deep copy dictionary.

I tried to return key value pairs from the display_transactions function but thought this would be easier for now.